+++ src/Application/Command/Category/Update.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Application\Command\Category;
 
 use Proximum\Vimeet\Domain\Model\Category;
@@ -15,33 +7,22 @@ use Proximum\Vimeet\Domain\Model\Event;
 
 class Update
 {
-    /**
-     * @var Category
-     */
+    /** @var Category */
     public $category;
 
-    /**
-     * @var Event
-     */
+    /** @var Event */
     public $event;
 
-    /**
-     * @var array
-     */
+    /** @var array */
     public $translations = [];
 
-    /**
-     * @var array
-     */
+    /** @var array */
     public $types = [];
 
-    /**
-     * @param Category $category
-     */
     public function __construct(Category $category)
     {
         $this->category = $category;
-        $this->event    = $category->getEvent();
+        $this->event = $category->getEvent();
 
         foreach ($category->getTranslations() as $translation) {
             $this->translations[$translation->getLocale()] = [