+++ src/Application/Command/Tip/Event/Create.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Application\Command\Tip\Event;
 
 use Proximum\Vimeet\Domain\Model\Event;
@@ -18,9 +10,6 @@ class Create extends AbstractEventTip
     /** @var Event */
     public $event;
 
-    /**
-     * @param Event $event
-     */
     public function __construct(Event $event)
     {
         $this->event = $event;