+++ src/Application/Command/Tip/Event/Update.php
@@ -1,13 +1,5 @@
 <?php
 
-/*
- * This file is part of the Proximum Vimeet project.
- *
- * Copyright (C) Proximum
- *
- * @author Elao <contact@elao.com>
- */
-
 namespace Proximum\Vimeet\Application\Command\Tip\Event;
 
 use Proximum\Vimeet\Domain\Model\Tip\Tip;
@@ -17,9 +9,6 @@ class Update extends AbstractEventTip
     /** @var Tip */
     public $tip;
 
-    /**
-     * @param Tip $tip
-     */
     public function __construct(Tip $tip)
     {
         $this->tip = $tip;
@@ -42,10 +31,14 @@ class Update extends AbstractEventTip
         $this->conditionIsCompleteSheet = $tip->hasConditionCompleteSheet();
         $this->conditionIsPhoneConfirmed = $tip->hasConditionPhoneConfirmed();
 
-        foreach ($tip->getTranslations() as $translation) {
-            $this->translations[$translation->getLocale()] = [
-                'title' => $translation->getTitle(),
-                'content' => $translation->getContent(),
+        foreach ($tip->getEvent()->getLocales() as $locale) {
+            $translation = $tip->getTranslation($locale);
+            $translationTitle = $translation ? $translation->getTitle() : '';
+            $translationContent = $translation ? $translation->getContent() : '';
+
+            $this->translations[$locale] = [
+                'title' => $translationTitle,
+                'content' => $translationContent,
             ];
         }
     }