+++ src/Ui/Bundle/EventBundle/Resources/assets/js/components/VideoConference/Settings.js
@@ -6,7 +6,8 @@ const WEBAUDIO_ANALYZER_SMOOTHING_TIME = 0.8;
  */
 function Settings(
     videoSettingsContainer,
-    settingsValidateCallback
+    settingsValidateCallback,
+    enableInvisibleMode
 ) {
     this.videoSettingsContainer = videoSettingsContainer;
     this.settingsValidateCallback = settingsValidateCallback;
@@ -16,6 +17,12 @@ function Settings(
     this.requestPermissionModal = this.videoSettingsContainer.querySelector('#visio-request-permission');
     this.settingsModal = this.videoSettingsContainer.querySelector('#visio-settings');
     this.requestPermissionErrorContainer = this.videoSettingsContainer.querySelector('#visio-request-permission-error');
+    this.invisibleModeOption = this.videoSettingsContainer.querySelector('#invisible-mode-option');
+    this.enableInvisibleMode = enableInvisibleMode;
+
+    if (enableInvisibleMode) {
+        showElement(this.invisibleModeOption);
+    }
 
     this.audioDeviceId = null;
     this.videoDeviceId = null;
@@ -298,7 +305,8 @@ Settings.prototype.prepareEventListener = function () {
 
     this.validateSettingsButton.addEventListener('click', (event) => {
         this.closeSettings();
-        this.settingsValidateCallback();
+        const invisibleMode = this.enableInvisibleMode && this.invisibleModeOption.querySelector('input[type=checkbox]').checked;
+        this.settingsValidateCallback(invisibleMode);
     });
 
     // Event dispatched when a device is plugged or unplugged from the computer/device of the user.